bring development up-to-date with master (#22)
* More OSX work; pgsql8.4.22 compiled for OS X, updated build instructions

* Included Java 6 JDK (Oracle license allows redistribution)

* Updated OSX.md, with emphasis on case sensitive file system

* Updated instructions for building from GitHub

* Updated, typos.

* Check for case-sensitive file system on OS X

* Changed to generic MacOSX.sdk, for when Xcode changes next month

* Change abort to warning on case insensitive experiments

* Fix pull of remote repostory

* Added dicom3tools (#26)

These tools are used to compile the FdfToDcm program (FDF to Dicom)
in the src/dicom_fdf category. Libraries for both Linux and MacOS
are present.

* Added missing Linux rpms (#28)

* Added dicom3tools

These tools are used to compile the FdfToDcm program (FDF to Dicom)
in the src/dicom_fdf category. Libraries for both Linux and MacOS
are present.

* Added missing Linux rpms

The install process gave errors due to missing rpm packages not
being installed. (tcl and tk). ALso added third party html to pdf
conversion program.

* Development (#21)

* Development (#8)

* Merge master in development (#1)

* Fixed so buildovj runs on Ubuntu (#5)

* Improvements to the buildovj script.

The ssh clone uses git@github.com instead of username@github.com.
Also, buildovj can now specify the branch (e.g., master or development)
to build. The default is to use a https and to build the master branch.

* Buildovj can be called from any directory on MacOS

* vagrant builds for centos7 and ubuntu14.04

* centos6 too

* fix build so it works on both centos6 and trusty

* updated vm docs

* docs update

* links and better VM instructions

* separate VM requirements and build steps

* clarify

* clarify, dont require OpenVnmrJ on build machine

* indentation

* added license headers, deliberately copying existing boilerplate to allow updates to future versions of the apache license

* Vagrant build cleanup (#14)

* Fully automatic build and VM setup from scratch.

Build and install tested on CentOS6 and Ubuntu14. Hopefully the start
of a CI infrastructure.  The three scripts here are:
- build_box.sh     builds only a vagrant box for a particular OS
- build_vm.sh      builds OpenVnmrJ on a box created by build_box.sh
- build_release.sh builds OpenVnmrJ source and calls packaging scripts
                   to create dvd images (work in progress).
added xenial,centos6

* added builds for macos10.10 and centos7, renamed ubuntu targets

* More build script improvements (#16)

* Fully automatic build and VM setup from scratch.

Build and install tested on CentOS6 and Ubuntu14. Hopefully the start
of a CI infrastructure.  The three scripts here are:
- build_box.sh     builds only a vagrant box for a particular OS
- build_vm.sh      builds OpenVnmrJ on a box created by build_box.sh
- build_release.sh builds OpenVnmrJ source and calls packaging scripts
                   to create dvd images (work in progress).
added xenial,centos6

* added builds for macos10.10 and centos7, renamed ubuntu targets

* build scripts with status msgs and error checking

* change default verbosity to "info"

* vagrant files for macos.

*obviously* only runnable on a vm on top of macos,
for licensing reasons.
also a small fix for the ubuntu14.14 box /etc/hosts

* remove cut/paste "typo", pass ovjTools branch to vm build (#17)

* Added dicom3tools (#23)

These tools are used to compile the FdfToDcm program (FDF to Dicom)
in the src/dicom_fdf category. Libraries for both Linux and MacOS
are present.

* Revert "Added dicom3tools (#23)" (#24)

This reverts commit 18b878e05b099d7bd91902ecc92ce1eef4b5a0d1.

* Dicom (#25)

* More OSX work; pgsql8.4.22 compiled for OS X, updated build instructions

* Included Java 6 JDK (Oracle license allows redistribution)

* Updated OSX.md, with emphasis on case sensitive file system

* Updated instructions for building from GitHub

* Updated, typos.

* Check for case-sensitive file system on OS X

* Changed to generic MacOSX.sdk, for when Xcode changes next month

* Change abort to warning on case insensitive experiments

* Fix pull of remote repostory

* Added dicom3tools

These tools are used to compile the FdfToDcm program (FDF to Dicom)
in the src/dicom_fdf category. Libraries for both Linux and MacOS
are present.

* Linux rpms  (#27)

* More OSX work; pgsql8.4.22 compiled for OS X, updated build instructions

* Included Java 6 JDK (Oracle license allows redistribution)

* Updated OSX.md, with emphasis on case sensitive file system

* Updated instructions for building from GitHub

* Updated, typos.

* Check for case-sensitive file system on OS X

* Changed to generic MacOSX.sdk, for when Xcode changes next month

* Change abort to warning on case insensitive experiments

* Fix pull of remote repostory

* Added dicom3tools

These tools are used to compile the FdfToDcm program (FDF to Dicom)
in the src/dicom_fdf category. Libraries for both Linux and MacOS
are present.

* Added missing Linux rpms

The install process gave errors due to missing rpm packages not
being installed. (tcl and tk). ALso added third party html to pdf
conversion program.

* cleanup packaging step (#20)